moved analysis files to src subdirectory
