+++ src/analysis/analyze_stack_depth.py
