+++ src/analysis/generate_conf_stack_depth.py
